Rename edge scr package to src
