+++ edge/src/mcc_reader.py
